Make UC1701 display conditional
Previously the UC1701 display was conditional on #define ARDUBOY2_SAMD
whereas now there is a separate #define ARDUBOY_UC1701 to control which
display (UC1701-based LCD or SSD1306-based OLED) is compiled/macro'd in.

+++ src/Arduboy2.cpp
@@ -337,7 +337,7 @@ unsigned long Arduboy2Base::generateRandomSeed()
 
   power_adc_disable(); // ADC off
 #else
-	seed = analogRead(A4) * ~micros() + micros();
+	seed = (unsigned long)analogRead(A4) * ~micros() + micros();
 #endif // ndef ARDUBOY_SAMD
   return seed;
 }
